+++ src/class/GameUtils.js
