+++ src/class/index.js
@@ -2,6 +2,6 @@ export { default as RacingGame } from './RacingGame.js';
 export { default as Cars } from './Cars.js';
 export { default as Car } from './Car.js';
 
-export { default as UserInterface } from './Interface.js';
+export { default as UserInterface } from './UserInterface.js';
 export { default as Validator } from './Validator.js';
-export { default as GameUtils } from './MyUtils.js';
+export { default as GameUtils } from './GameUtils.js';
